+++ src/Ydb.Sdk/src/Properties/AssemblyInfo.cs
@@ -1,6 +1,7 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("EntityFrameworkCore.Ydb")]
+[assembly: InternalsVisibleTo("EntityFrameworkCore.Ydb.FunctionalTests")]
 [assembly: InternalsVisibleTo("Ydb.Sdk.Ado.Benchmarks")]
 [assembly: InternalsVisibleTo("Ydb.Sdk.Ado.Tests")]
 [assembly: InternalsVisibleTo("Ydb.Sdk.Topic.Tests")]