+++ src/Ydb.Sdk/src/Topic/Reader/Reader.cs
@@ -168,7 +168,7 @@ internal class Reader<TValue> : IReader<TValue>
                 var initException = YdbException.FromServer(receivedInitMessage.Status, receivedInitMessage.Issues);
                 var statusMessage = initException.Message;
 
-                if (YdbRetryPolicy.IdempotenceDefault.GetNextDelay(initException, attempt: 0) is not null)
+                if (YdbRetryPolicy.IdempotenceDefault.GetNextDelay(initException, attempt: 1) is not null)
                 {
                     _logger.LogError("Reader initialization failed to start. {StatusMessage}", statusMessage);